fix(CRA-scripts): disable `fullySpecified` resolution for `.mjs` files
[Resolve configuration](https://webpack.js.org/configuration/module/#resolvefullyspecified):

When enabled, you should provide the file extension
when importing a module in .mjs files
or any other .js files when their nearest parent package.json file
contains a "type" field with a value of "module",
otherwise webpack would fail the compiling with a Module not found error.

And webpack won't resolve directories with filenames defined in the resolve.mainFiles,
you have to specify the filename yourself.